Removes some shitty code
