+++ yogstation.dme
@@ -742,7 +742,6 @@
 #include "code\game\machinery\requests_console.dm"
 #include "code\game\machinery\scan_gate.dm"
 #include "code\game\machinery\sci_bombardment.dm"
-#include "code\game\machinery\sci_probe.dm"
 #include "code\game\machinery\shieldgen.dm"
 #include "code\game\machinery\Sleeper.dm"
 #include "code\game\machinery\slotmachine.dm"
